+++ src/engine/Scene.ts
@@ -1,6 +1,6 @@
 import { UIActor } from './UIActor';
 import { Physics } from './Physics';
-import { InitializeEvent, ActivateEvent, DeactivateEvent, PreUpdateEvent, PostUpdateEvent, 
+import { InitializeEvent, ActivateEvent, DeactivateEvent, PreUpdateEvent, PostUpdateEvent,
    PreDrawEvent, PostDrawEvent, PreDebugDrawEvent, PostDebugDrawEvent, GameEvent } from './Events';
 import { Logger } from './Util/Log';
 import { Timer } from './Timer';
@@ -16,10 +16,10 @@ import { Class } from './Class';
 import * as Util from './Util/Util';
 import * as Events from './Events';
 import * as ActorUtils from './Util/Actors';
-
+import { obsolete } from './Util/Decorators';
 /**
- * [[Actor|Actors]] are composed together into groupings called Scenes in 
- * Excalibur. The metaphor models the same idea behind real world 
+ * [[Actor|Actors]] are composed together into groupings called Scenes in
+ * Excalibur. The metaphor models the same idea behind real world
  * actors in a scene. Only actors in scenes will be updated and drawn.
  *
  * Typical usages of a scene include: levels, menus, loading screens, etc.
@@ -41,7 +41,7 @@ export class Scene extends Class {
    /**
     * The actors in the current scene
     */
-   public children: Actor[] = [];
+   public actors: Actor[] = [];
 
    /**
     * The [[TileMap]]s in the scene, if any
@@ -92,7 +92,7 @@ export class Scene extends Class {
    public on(eventName: Events.predraw, handler: (event?: PreDrawEvent) => void): void;
    public on(eventName: Events.postdraw, handler: (event?: PostDrawEvent) => void): void;
    public on(eventName: Events.predebugdraw, handler: (event?: PreDebugDrawEvent) => void): void;
-   public on(eventName: Events.postdebugdraw, handler: (event?: PostDebugDrawEvent) => void): void;   
+   public on(eventName: Events.postdebugdraw, handler: (event?: PostDebugDrawEvent) => void): void;
    public on(eventName: string, handler: (event?: GameEvent<any>) => void): void;
    public on(eventName: string, handler: (event?: GameEvent<any>) => void): void {
       super.on(eventName, handler);
@@ -134,7 +134,7 @@ export class Scene extends Class {
     * Initializes actors in the scene
     */
    private _initializeChildren(): void {
-      for (var child of this.children) {
+      for (var child of this.actors) {
          child._initialize(this.engine);
       }
    }
@@ -147,7 +147,7 @@ export class Scene extends Class {
    }
 
    /**
-    * Initializes the scene before the first update, meant to be called by engine not by users of 
+    * Initializes the scene before the first update, meant to be called by engine not by users of
     * Excalibur
     * @internal
     */
@@ -180,10 +180,9 @@ export class Scene extends Class {
       this._cancelQueue.length = 0;
 
       // Cycle through timers updating timers
-      this._timers = this._timers.filter(timer => {
+      for (var timer of this._timers) {
          timer.update(delta);
-         return !timer.complete;
-      });
+      };
 
       // Cycle through actors updating UI actors
       for (i = 0, len = this.uiActors.length; i < len; i++) {
@@ -196,8 +195,8 @@ export class Scene extends Class {
       }
 
       // Cycle through actors updating actors
-      for (i = 0, len = this.children.length; i < len; i++) {         
-         this.children[i].update(engine, delta);
+      for (i = 0, len = this.actors.length; i < len; i++) {
+         this.actors[i].update(engine, delta);
       }
 
       this._collectActorStats(engine);
@@ -205,8 +204,8 @@ export class Scene extends Class {
       // Run the broadphase and narrowphase
       if (this._broadphase && Physics.enabled) {
          var beforeBroadphase = Date.now();
-         this._broadphase.update(this.children, delta);
-         var pairs = this._broadphase.broadphase(this.children, delta, engine.stats.currFrame);
+         this._broadphase.update(this.actors, delta);
+         var pairs = this._broadphase.broadphase(this.actors, delta, engine.stats.currFrame);
          var afterBroadphase = Date.now();
 
          var beforeNarrowphase = Date.now();
@@ -228,14 +227,14 @@ export class Scene extends Class {
       var actorIndex: number;
 
       for (i = 0, len = this._killQueue.length; i < len; i++) {
-         actorIndex = this.children.indexOf(this._killQueue[i]);
+         actorIndex = this.actors.indexOf(this._killQueue[i]);
          if (actorIndex > -1) {
             this._sortedDrawingTree.removeByComparable(this._killQueue[i]);
-            this.children.splice(actorIndex, 1);
+            this.actors.splice(actorIndex, 1);
          }
       }
       engine.stats.currFrame.actors.killed = this._killQueue.length;
-      this._killQueue.length = 0;      
+      this._killQueue.length = 0;
 
       this.emit('postupdate', new PostUpdateEvent(engine, delta, this));
    }
@@ -304,8 +303,8 @@ export class Scene extends Class {
          this.tileMaps[i].debugDraw(ctx);
       }
 
-      for (i = 0, len = this.children.length; i < len; i++) {
-         this.children[i].debugDraw(ctx);
+      for (i = 0, len = this.actors.length; i < len; i++) {
+         this.actors[i].debugDraw(ctx);
       }
 
       this._broadphase.debugDraw(ctx, 20);
@@ -318,7 +317,7 @@ export class Scene extends Class {
     * Checks whether an actor is contained in this scene or not
     */
    public contains(actor: Actor): boolean {
-      return this.children.indexOf(actor) > -1;
+      return this.actors.indexOf(actor) > -1;
    }
 
    /**
@@ -354,7 +353,7 @@ export class Scene extends Class {
          return;
       }
       if (entity instanceof Actor) {
-         if (!Util.contains(this.children, entity)) {
+         if (!Util.contains(this.actors, entity)) {
             this._addChild(entity);
             this._sortedDrawingTree.add(entity);
          }
@@ -375,7 +374,7 @@ export class Scene extends Class {
 
    /**
     * Removes a [[Timer]] from the current scene, it will no longer be updated.
-    * @param timer  The timer to remove to the current scene.       
+    * @param timer  The timer to remove to the current scene.
     */
    public remove(timer: Timer): void;
 
@@ -387,7 +386,7 @@ export class Scene extends Class {
 
    /**
     * Removes an actor from the scene, it will no longer be drawn or updated.
-    * @param actor  The actor to remove from the current scene.      
+    * @param actor  The actor to remove from the current scene.
     */
    public remove(actor: Actor): void;
 
@@ -436,12 +435,12 @@ export class Scene extends Class {
    }
 
    /**
-    * Adds an actor to the scene, once this is done the actor will be drawn and updated.       
+    * Adds an actor to the scene, once this is done the actor will be drawn and updated.
     */
    protected _addChild(actor: Actor) {
       this._broadphase.track(actor.body);
       actor.scene = this;
-      this.children.push(actor);
+      this.actors.push(actor);
       this._sortedDrawingTree.add(actor);
       actor.parent = this.actor;
    }
@@ -483,7 +482,7 @@ export class Scene extends Class {
    }
 
    /**
-    * Removes a [[Timer]] from the scene. 
+    * Removes a [[Timer]] from the scene.
     * @warning Can be dangerous, use [[cancelTimer]] instead
     * @param timer  The timer to remove
     */
@@ -508,7 +507,7 @@ export class Scene extends Class {
     * Tests whether a [[Timer]] is active in the scene
     */
    public isTimerActive(timer: Timer): boolean {
-      return (this._timers.indexOf(timer) > -1);
+      return (this._timers.indexOf(timer) > -1 && !timer.complete);
    }
 
    /**
@@ -563,7 +562,7 @@ export class Scene extends Class {
          engine.stats.currFrame.actors.ui++;
       }
 
-      for (var actor of this.children) {
+      for (var actor of this.actors) {
          engine.stats.currFrame.actors.alive++;
          for (var child of actor.children) {
             if (ActorUtils.isUIActor(child)) {
@@ -573,5 +572,14 @@ export class Scene extends Class {
             }
          }
       }
+   }
+
+   @obsolete({ alternateMethod: 'ex.Scene.actors', message: 'This method will be removed in the 0.13.0 release. '})
+   public set children(actors: Actor[]) {
+      this.actors = actors;
+   }
+
+   public get children() {
+      return this.actors;
    }
 }