+++ CHANGELOG.md
@@ -9,7 +9,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Breaking Changes
 ### Added
+- Added new hsl and hex format options in Color.toString(format) using rgb as the default to maintain backwards compatibility ([#852](https://github.com/excaliburjs/Excalibur/issues/852))
 ### Changed
+- Changed `ex.Scene.children` to be `ex.Scene.actors`, and updated unit test specs to reflect this change ([#796](https://github.com/excaliburjs/Excalibur/issues/796))
+- Added backgroundColor to engine options as part of Engine constructor ([#846](https://github.com/excaliburjs/Excalibur/issues/846))
 ### Deprecated
 ### Fixed
 
@@ -29,6 +32,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Camera zoom over time now returns a promise that resolves on completion ([#800](https://github.com/excaliburjs/Excalibur/issues/800))
 - Edge builds have more descriptive versions now containing build number and Git commit hash (e.g. `0.10.0-alpha.105#commit`) ([#777](https://github.com/excaliburjs/Excalibur/issues/777))
+- `Animation.loop` property now to set to `true` by default ([#583] (https://github.com/excaliburjs/Excalibur/issues/583))
 
 ### Fixed
 - Fixed camera zoom over time, before it did not work at all ([#800](https://github.com/excaliburjs/Excalibur/issues/800))